proof of concept: generics to remove duplicate code
zero performance impact, but requires go1.18+, and simply removes some
code duplication. since the code duplication was the original reason
this module was rejected for being upstreamed into the standard library,
it's possible that this may be a better way to go in the future,
especially once constraints.Ordered makes its way into the standard lib.
Of course, at that point, you'd probably want to make SearchComparable
or the like an exported function, and I suspect that will end up in the
standard library anyhow once the `ordered` debate gets figured out.

So for now just stashing in a separate branch to preserve the code.